+++ test/glob.test.js
@@ -4,7 +4,7 @@ var assert  = require("assert"),
     
     glob = require("../src/glob"),
     
-    compare = require("./lib/compare-files");
+    compare = require("./lib/compare.js");
 
 describe("/glob.js", function() {
     it("should be a function", function() {
