+++ test/issue-56.test.js
@@ -4,8 +4,8 @@ var assert = require("assert"),
     
     plugin = require("../src/plugin.js"),
 
-    compositions = require("./lib/compositions.js"),
-    compare      = require("./lib/compare-files.js");
+    exported = require("./lib/exported.js"),
+    compare  = require("./lib/compare.js");
 
 describe("/issues", function() {
     describe("/56", function() {
@@ -15,7 +15,7 @@ describe("/issues", function() {
                     { from : "./test/specimens/issues/56.css" }
             )
             .then((result) => {
-                assert.deepEqual(compositions(result), {
+                assert.deepEqual(exported(result).exports, {
                     "test/specimens/issues/56.css" : {
                         booga : "mc13e7db14_booga",
                         fooga : "mc13e7db14_booga mc13e7db14_fooga",
