+++ test/lib/exported.js
@@ -1,0 +1,4 @@
+"use strict";
+
+module.exports = (result) =>
+    result.messages.find((msg) => (msg.name === "modular-css")) || false;
